move knoweldge handlers to other file
